requirements.txt
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.3.1
defusedxml==0.6.0
Django==3.1.6
django-allauth==0.44.0
django-cors-headers==3.7.0
django-debug-toolbar==3.2
django-debug-toolbar-force==0.1.8
django-filter==2.4.0
django-nine==0.2.4
django-rest-auth==0.9.5
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
drf-yasg==1.20.0
idna==2.10
inflection==0.5.1
install==1.3.4
itypes==1.2.0
Jinja2==2.11.3
MarkupSafe==1.1.1
oauthlib==3.1.0
packaging==20.9
psycopg2==2.8.6
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.7
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
ruamel.yaml==0.16.12
six==1.15.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.3
